requirements.txt
firebase-admin == 4.3.0
django-heroku == 0.3.1
django == 3.1.1
gunicorn == 20.0.4
pandas == 1.1.1
scikit-learn == 0.23.2
scikit-surprise == 1.1.1
matplotlib == 3.3.1